Enable use of CLI web browsers like Lynx from inside console Vim
I'm pretty sure the plug-in can't tell the difference between
GUI and CLI web browsers, therefor I've included a specific list
of CLI web browsers (Lynx and Links for now) which should be
opened interactively in console Vim's alternate screen.

To enable one of these web browsers set the global variable
pyref_browser to the name or path of `lynx' or `links'.

Note that the following line:

	execute '!' . browser fnameescape(a:url)

Should really be:

	execute '!' . browser fnameescape(shellescape(a:url))

But unfortunately that doesn't work because Lynx gives a "Bad URL
scheme" error message and exits (because Vim quotes command-line
arguments using single-quotes, which apparently aren't removed
when the CLI web browser is started using the ! command?!)